pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "py-moodle"
dynamic = ["version"]
description = "Modern Moodle CLI tool"
readme = "README.md"
authors = [{ name = "Ernesto Serrano", email = "info@ernesto.es" }]
requires-python = ">=3.8"
dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
    "typer[all]",
    "rich==14.1.0",
]

[project.scripts]
py-moodle = "py_moodle.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "release-branch-semver"

[tool.black]
line-length = 88
exclude = '''
/(
    env
)/
'''

[tool.isort]
profile = "black"
skip = ["env"]

[tool.flake8]
ignore = ["E501", "W503"]
